refactor(BA-6057): remove legacy scheduler package superseded by Sokovan
- Delete src/ai/backend/manager/scheduler/ (FIFO/LIFO/DRF schedulers,
  agent selectors, predicates, utils) — all functionality has been ported
  to manager/sokovan/scheduler/.
- Move ScheduleType enum into manager/sokovan/scheduler/types.py and
  update its six importers.
- Remove dead settle_agent_alloc() and KernelAgentBinding usage from
  AgentRegistry (no callers since the Phase 3 JSONB write removal).
- Drop the orphaned backendai_scheduler_v10 and backendai_agentselector_v10
  entrypoint groups from BUILD and plugin entrypoint metadata.
- Update tests/unit/plugin/test_entrypoint.py fixtures, model row
  comments, and manager/CLAUDE.md Scheduler guardrail to point at the
  Sokovan location.
